feat: makes the jira worker run and wokers quit properly
**Infra Changes**:

separated the worker and sever. will need to change the Dockerfile
to use pm2-client, or use docker-compose.

**Code Changes**:

The WorkerPool can be kept as a single instance. And then one can
provide multiple types of Worker.
- ProducerConsumer
- Consumer

With the new workerpool package, the GithubWorkerPool can now be
removed and the worker has to become an interface. Have set the
stage for this changes.